+++ src/main/java/racing/domain/move/MoveStrategy.java
@@ -1,5 +1,6 @@
-package racing.strategy;
+package racing.domain.move;
 
+@FunctionalInterface
 public interface MoveStrategy {
     boolean move();
 }
